+++ CHANGELOG.md
@@ -93,6 +93,7 @@
 - Added ability to autoescape Flash messages [#11448](https://github.com/phalcon/cphalcon/issues/11448)
 - Fixed `Phalcon\Config::merge` for working with php7
 - Added ability to use custom delimiter for `Phalcon\Text::camelize` and `Phalcon\Text::uncamelize` [#10396](https://github.com/phalcon/cphalcon/issues/10396)
+- Added support of `CONNECT`, `TRACE` and `PURGE`  HTTP methods
 
 # [2.0.13](https://github.com/phalcon/cphalcon/releases/tag/phalcon-v2.0.13) (2016-05-19)
 - Restored `Phalcon\Text::camelize` behavior [#11767](https://github.com/phalcon/cphalcon/issues/11767)